This is the old relaxation branch that we used to work on. It "works" with the (albeit, wrong) assumptions made...
The assumption was that relaxing labels would be fine regardless. This is
not true as labels are relative to the start of the section the label resides
in. This means that pc-relative instructions that use labels should be fine
but that instructions using labels as absolute values won't work because
the linker will take the section and place is into memory at some absolute
address which might completely mess up the instruction used as the absolute
value might be outside of the range that instruction supports.